refactored test code to test folder, fixed data-vis tests, added empty modules, switched metadata and header order for IO
