+++ src/bands/Module.jl
@@ -1,0 +1,3 @@
+module BandsModule
+
+end
