--- src/data-vis/constants.jl
@@ -1,5 +1,0 @@
-# contains constants used in make-cnt-data and other data visualization files
-
-const nm = 1E-9
-const δ₀ = 0.142*nm
-const a = 0.246*nm
