+++ src/negf/Module.jl
@@ -1,0 +1,3 @@
+module NEGFModule
+
+end
